requirements.txt
# docker build

# torch~=2.1.2
# torchaudio~=2.1.2

transformers~=4.38.1
pydantic~=2.5.3
fastapi~=0.109.0
python-multipart~=0.0.6
numpy~=1.25.2
python-dotenv~=1.0.1
uvicorn~=0.27.1
librosa~=0.10.1

# 
# supabase~=2.4.0
# requests~=2.31.0
# soundfile~=0.12.1

# UI
# streamlit~=1.30.0
# audio-recorder-streamlit~=0.0.8

streamlit~=1.31.1
sounddevice~=0.4.6
soundfile~=0.12.1
requests~=2.31.0
supabase~=2.4.0
scipy~=1.11.2
torch~=2.2.1